feat(cli-client): full go-flags parser for upload
